Moved detector_px_to_3D_kspace to vector_utils file
The detector_px_to_3D_kspace has been moved to vector_utils file, to make code clearer and following the calculate_cartesian_coordinates example.
